Phone number is unique for worker

+++ src/main/java/org/mts/internship/service/WorkerService.java
@@ -7,6 +7,7 @@ import org.mts.internship.dto.WorkerDto;
 import org.mts.internship.entity.Department;
 import org.mts.internship.entity.Worker;
 import org.mts.internship.exception.DepartmentNotFoundException;
+import org.mts.internship.exception.WorkerExistsByPhoneNumberException;
 import org.mts.internship.exception.WorkerNotFoundException;
 import org.mts.internship.mapper.WorkerMapper;
 import org.mts.internship.repository.DepartmentRepository;
@@ -24,49 +25,56 @@ public class WorkerService {
 
     private final WorkerRepository workerRepository;
     private final DepartmentRepository departmentRepository;
+    private final WorkerMapper workerMapper;
 
     public Set<WorkerDto> getAllWorkers() {
         List<Worker> workers = (List<Worker>) workerRepository.findAll();
         return workers.stream()
-                .map(WorkerMapper::mapToDto)
+                .map(workerMapper::mapToDto)
                 .collect(Collectors.toSet());
     }
 
     public WorkerDto getById(long id) {
         Worker worker = workerRepository.findById(id)
                 .orElseThrow(() -> new WorkerNotFoundException(id));
-        return WorkerMapper.mapToDto(worker);
+        return workerMapper.mapToDto(worker);
     }
 
     @Transactional
     public WorkerDto createWorker(CreateWorkerRequest request) {
         Department department = departmentRepository.findByName(request.getDepartmentName())
                 .orElseThrow(() -> new DepartmentNotFoundException(request.getDepartmentName()));
-        Worker worker = new Worker(request.getFirstName(), request.getLastName(), request.getPhoneNumber(),
-                department);
+        if(workerRepository.existsByPhoneNumber(request.getPhoneNumber())){
+            throw new WorkerExistsByPhoneNumberException(request.getPhoneNumber());
+        }
+        Worker worker = new Worker()
+                .setFirstName(request.getFirstName())
+                .setLastName(request.getLastName())
+                .setPhoneNumber(request.getPhoneNumber())
+                .setDepartment(department);
 
         workerRepository.save(worker);
-        return WorkerMapper.mapToDto(worker);
+        return workerMapper.mapToDto(worker);
     }
 
     @Transactional
-    public WorkerDto updateWorker(UpdateWorkerRequest request) {
-        Worker worker = workerRepository.findById(request.getId())
-                .orElseThrow(() -> new WorkerNotFoundException(request.getId()));
+    public WorkerDto updateWorker(long id, UpdateWorkerRequest request) {
+        Worker worker = workerRepository.findById(id)
+                .orElseThrow(() -> new WorkerNotFoundException(id));
 
         if (request.getFirstName() != null) worker.setFirstName(request.getFirstName());
         if (request.getLastName() != null) worker.setLastName(request.getLastName());
         if (request.getPhoneNumber() != null) worker.setPhoneNumber(request.getPhoneNumber());
 
         workerRepository.save(worker);
-        return WorkerMapper.mapToDto(worker);
+        return workerMapper.mapToDto(worker);
     }
 
     @Transactional
-    public WorkerDto deleteWorker(long id) {
-        Worker worker = workerRepository.findById(id)
-                .orElseThrow(() -> new WorkerNotFoundException(id));
+    public void deleteWorker(long id) {
+        if(!workerRepository.existsById(id)){
+            throw new WorkerNotFoundException(id);
+        }
         workerRepository.deleteById(id);
-        return WorkerMapper.mapToDto(worker);
     }
 }
